fix: handle no more code case, refactor and improve performance
